Change all variables storing get_millisec() results to uint32_t
- Update device/life_device.h, device/valve.h, device/device.h variable types
- Update IoT/iot_altivar.h time variables to uint32_t
- Update tcp_client.h and platform-specific headers (l_tcp_client.h, w_tcp_client.h)
- Update tcp_cmctr headers and implementations
- Update bus_coupler_io.h, params_recipe_manager.h time variables
- Update device.cpp timer methods implementations
- Update tcp_client constructors in tcp_client.cpp and platform-specific files
- Update modbus_client and iot_altivar constructors
- Update static variables in wago_w750.cpp, PAC_err.cpp, g_device.cpp
- Update local variables in l_tcp_cmctr.cpp, uni_bus_coupler_io.cpp, tcp_cmctr.cpp
- Update PAC_dev_lua.cpp tolua function
- Add #include <cstdint> to iot_altivar.h
- All 610 tests pass

+++ PAC/common/PAC_dev_lua.cpp
@@ -8730,7 +8730,7 @@ static int tolua_PAC_dev_get_millisec00(lua_State* tolua_S)
 #endif
  {
   {
-   unsigned long tolua_ret = (unsigned long)  get_millisec();
+   uint32_t tolua_ret = (uint32_t)  get_millisec();
    tolua_pushnumber(tolua_S,(lua_Number)tolua_ret);
   }
  }
